talk-to-task/requirements.txt
#API
fastapi
uvicorn

#Env variables
python-dotenv

#AI

langchain
langchain-openai
langchain-community
langgraph